test: add unit tests for needs-decision snapshot body
