+++ .github/scripts/needs-decision-snapshot.ts
@@ -8,15 +8,8 @@
  * - Low-noise: updates one existing issue body (or creates it once if missing).
  */
 
-type GitHubSearchItem = {
-  number: number;
-  title: string;
-  html_url: string;
-  user?: { login?: string };
-  updated_at?: string;
-  created_at?: string;
-  pull_request?: unknown;
-};
+import { buildBody, LABEL, SNAPSHOT_TITLE, splitPrsAndIssues } from './needs-decision-snapshot-lib.js';
+import type { GitHubSearchItem } from './needs-decision-snapshot-lib.js';
 
 type GitHubSearchResponse = {
   total_count: number;
@@ -31,9 +24,6 @@ type GitHubIssue = {
   html_url: string;
   body: string | null;
 };
-
-const SNAPSHOT_TITLE = 'Needs-decision snapshot (automated)';
-const LABEL = 'needs-decision';
 
 function mustEnv(name: string): string {
   const v = process.env[name];
@@ -54,18 +44,6 @@ function toBool(v: string | undefined, defaultValue = false): boolean {
   if (['1', 'true', 'yes', 'y', 'on'].includes(s)) return true;
   if (['0', 'false', 'no', 'n', 'off'].includes(s)) return false;
   return defaultValue;
-}
-
-function fmtDate(dateLike?: string): string {
-  if (!dateLike) return '';
-  const d = new Date(dateLike);
-  if (Number.isNaN(d.getTime())) return '';
-  return d.toISOString().slice(0, 10);
-}
-
-function fmtIsoMinute(d: Date): string {
-  const iso = d.toISOString();
-  return iso.slice(0, 16).replace('T', ' ') + ' UTC';
 }
 
 async function ghFetch<T>(path: string, init?: RequestInit): Promise<T> {
@@ -107,78 +85,6 @@ async function searchAllIssues(query: string, maxPages = 10): Promise<GitHubSear
   }
 
   return items;
-}
-
-function splitPrsAndIssues(items: GitHubSearchItem[]): { prs: GitHubSearchItem[]; issues: GitHubSearchItem[] } {
-  const prs: GitHubSearchItem[] = [];
-  const issues: GitHubSearchItem[] = [];
-
-  for (const it of items) {
-    if (it.pull_request) prs.push(it);
-    else issues.push(it);
-  }
-
-  return { prs, issues };
-}
-
-function renderList(items: GitHubSearchItem[], limit: number): string {
-  if (items.length === 0) return '_None._\n';
-
-  const shown = items.slice(0, limit);
-  const lines = shown.map((it) => {
-    const who = it.user?.login ? `@${it.user.login}` : 'unknown';
-    const updated = fmtDate(it.updated_at);
-    const suffix = updated ? ` (updated ${updated})` : '';
-    return `- [#${it.number}](${it.html_url}) ${it.title} — ${who}${suffix}`;
-  });
-
-  const extra = items.length - shown.length;
-  if (extra > 0) {
-    lines.push(`- …and ${extra} more`);
-  }
-
-  return lines.join('\n') + '\n';
-}
-
-function buildBody(params: {
-  repoFull: string;
-  runUrl?: string;
-  generatedAt: Date;
-  prs: GitHubSearchItem[];
-  issues: GitHubSearchItem[];
-}): string {
-  const { repoFull, runUrl, generatedAt, prs, issues } = params;
-
-  const total = prs.length + issues.length;
-  const qUrl = `https://github.com/${repoFull}/issues?q=is%3Aopen+label%3A${encodeURIComponent(LABEL)}`;
-
-  const runLine = runUrl ? ` | [workflow run](${runUrl})` : '';
-
-  const marker = '<!-- needs-decision-snapshot: do-not-edit -->';
-
-  // Keep noise low: compact, scannable.
-  return [
-    marker,
-    `# Needs-decision snapshot`,
-    '',
-    `Open items labeled \`${LABEL}\`: **${total}** (PRs: **${prs.length}**, issues: **${issues.length}**)`,
-    '',
-    `Last updated: **${fmtIsoMinute(generatedAt)}**${runLine}`,
-    '',
-    `Query: ${qUrl}`,
-    '',
-    '## Pull requests',
-    '',
-    renderList(prs, 50).trimEnd(),
-    '',
-    '## Issues',
-    '',
-    renderList(issues, 50).trimEnd(),
-    '',
-    '---',
-    '',
-    'This issue is maintained automatically by a scheduled GitHub Actions workflow.',
-  ].join('\n');
 }
 
 async function getIssue(owner: string, repo: string, issueNumber: number): Promise<GitHubIssue> {
